+++ .storybook/config.js
@@ -1,7 +1,6 @@
 import { configure } from '@storybook/react';
 
 import 'psychic-ui/dist/psychic-min.css';
-import '../src/style.css';
 
 function loadStories() {
   require('../stories');