+++ stories/index.js
@@ -3,8 +3,10 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 
 import { HTML, Text, Editor } from '../src/index';
+
 import Code from '../src/templates/code.js';
 import Product from '../src/templates/product.js';
+import Landing from '../src/templates/landing.js';
 
 storiesOf('Text', module)
   .add('basic text', () => <Text value="Hello I am text!"></Text>)
@@ -15,6 +17,10 @@ storiesOf('HTML', module)
 
 storiesOf('Editor', module)
   .add('hello world javascript', () => {
+    window.extra = (string) => {
+      return `!${string}!`
+    }
+
     var options = {
       title: "Hello World",
       subtitle: "This is a hello world example",
@@ -61,10 +67,36 @@ storiesOf('Product Template', module)
     var options = {
         title: "Steno",
         description: "A simple SSH shortcut menu for OSX",
-        sourceCodeLink: 'https://github.com/gabrielcsapo/steno',
-        downloadLink: 'https://github.com/gabrielcsapo/steno/releases',
+        links: {
+          Source: 'https://github.com/gabrielcsapo/steno',
+          Download: 'https://github.com/gabrielcsapo/steno/releases',
+        },
         icon: require('../test/fixtures/steno.png'),
         demoImage: require('../test/fixtures/example.gif')
     }
     return <Product {...options}/>
+  });
+
+storiesOf('Landing Template', module)
+  .add('Example Landing Template', () => {
+    const options = {
+        title: 'Steno',
+        nav: {
+          Docs: 'http://gabrielcsapo.com/steno'
+        },
+        body: `
+          <div style="text-align:center;">
+            <h4 style="font-weight:100">A simple SSH shortcut menu for OSX</h4>
+            <img class="responsive" src="${require('../test/fixtures/example.gif')}"/>
+          </div>
+        `,
+        options: {
+          width: '50%'
+        },
+        footer: {
+          author: 'Gabriel J. Csapo',
+          website: 'http://www.gabrielcsapo.com'
+        }
+    };
+    return <Landing {...options}/>
   });